Fix navigation portlet upgrade step: increase version numbers properly.
