+++ ftw/book/upgrades/to2200.py
@@ -2,10 +2,6 @@ from Acquisition import aq_parent, aq_inner
 from ftw.upgrade import ProgressLogger
 from ftw.upgrade import UpgradeStep
 import logging
-from ftw.book.eventhandler import add_navigation_portlet
-from plone.portlets.interfaces import IPortletAssignmentMapping
-from plone.portlets.interfaces import IPortletManager
-from zope.component import getUtility, getMultiAdapter
 
 
 LOG = logging.getLogger('ftw.book.upgrades')
@@ -49,21 +45,3 @@ class MakeBlocksSearchable(UpgradeStep):
                 parent = aq_parent(aq_inner(obj))
                 parent.reindexObject(idxs=['SearchableText'])
                 step()
-
-
-class SetBookPortlest(UpgradeStep):
-
-    def __call__(self):
-        catalog = self.getToolByName('portal_catalog')
-        brains = catalog(portal_type='Book')
-        for brain in brains:
-            book = brain.getObject()
-            manager = getUtility(IPortletManager, name='plone.leftcolumn')
-            mapping = getMultiAdapter((book, manager),
-                                  IPortletAssignmentMapping).__of__(book)
-            path = '/'.join(book.getPhysicalPath())[
-                len(book.portal_url.getPortalPath()):]
-            navi = mapping.get('navigation')
-            if navi and navi.root == path:
-                continue
-            add_navigation_portlet(book, None)
